New requirement on guest is implemented

+++ lib/widgets/sign_up_dialog.dart
@@ -1,7 +1,8 @@
+import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import '../presentation/sign_in_screen/sign_in_screen.dart';
-import '../routes/routes.dart';
+import '../presentation/welcome_screen/welcome_screen.dart';
+import '../service/api/auth_service.dart';
 import '../service/device_service.dart';
 import '../utils/scaffold_messenger_utils.dart';
 
@@ -20,9 +21,39 @@ void showSignUpDialog(BuildContext context,bool timeOver ) async {
           TextButton(
             onPressed: () async {
               if(timeOver){
-                deviceService.removeDeviceId();
-                Navigator.pop(context, true);
-                Get.offAll(SignInScreen());
+                DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+                AndroidDeviceInfo androidInfo = await deviceInfo.androidInfo;
+
+                String? deviceId = androidInfo.id; // Fetch unique device ID
+
+                if (deviceId != null) {
+                  print('Device ID: $deviceId'); // Log or use it as needed
+
+                  // Show loading indicator
+                  showDialog(
+                    context: context,
+                    barrierDismissible: false, // Prevent dismissing the dialog by tapping outside
+                    builder: (context) => const Center(
+                      child: CircularProgressIndicator(),
+                    ),
+                  );
+
+                  try {
+                    // Call guestUser method
+                    await AuthService().guestUser(deviceId: deviceId);
+                    showScaffoldMessage(context, "Guest access finished, please sign up to continue.");
+                    Navigator.pop(context, true);
+                    Get.offAll(WelcomeScreen());
+                  } catch (e) {
+                    showScaffoldMessage(context, "Guest access finished, please sign up to continue.");
+                    Navigator.pop(context, true);
+                    deviceService.removeDeviceId();
+                    Get.offAll(WelcomeScreen());
+
+                  }
+                } else {
+                  print('Failed to retrieve device ID');
+                }
               }
               else{
                 Navigator.pop(context, true);
@@ -39,8 +70,39 @@ void showSignUpDialog(BuildContext context,bool timeOver ) async {
 
   if (result == null) {
     if(timeOver){
-      deviceService.removeDeviceId();
-      Get.offAll(SignInScreen());
+
+      DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+      AndroidDeviceInfo androidInfo = await deviceInfo.androidInfo;
+      String? deviceId = androidInfo.id; // Fetch unique device ID
+
+      if (deviceId != null) {
+        print('Device ID: $deviceId'); // Log or use it as needed
+
+        // Show loading indicator
+        showDialog(
+          context: context,
+          barrierDismissible: false, // Prevent dismissing the dialog by tapping outside
+          builder: (context) => const Center(
+            child: CircularProgressIndicator(),
+          ),
+        );
+
+        try {
+          // Call guestUser method
+          await AuthService().guestUser(deviceId: deviceId);
+          showScaffoldMessage(context, "Free guest access finished, you will see frequent ad, for ad free experience try sign in");
+          Navigator.pop(context, true);
+          Get.offAll(WelcomeScreen());
+        } catch (e) {
+          showScaffoldMessage(context, "Free guest access finished, you will see frequent ad, for ad free experience try sign in");
+          Navigator.pop(context, true);
+          deviceService.removeDeviceId();
+          Get.offAll(WelcomeScreen());
+
+        }
+      } else {
+        print('Failed to retrieve device ID');
+      }
     }
